refactored footer into shared footer
