+++ src/applications/health-care-questionnaire/questionnaire/components/confirmation-page-footer/ConfirmationPageFooter.jsx
