+++ src/applications/health-care-questionnaire/shared/components/footer/index.js
@@ -1,7 +1,4 @@
 import GetHelpFooter from './GetHelpFooter';
 import NeedHelpSmall from './NeedHelpSmall';
 
-export default {
-  footer: GetHelpFooter,
-  review: NeedHelpSmall,
-};
+export { GetHelpFooter, NeedHelpSmall };
