+++ src/applications/health-care-questionnaire/shared/components/footer/tests/utils.js
@@ -1,0 +1,20 @@
+const createFakeFooterStore = (clinic = {}, clinicFriendlyName = '') => {
+  return {
+    getState: () => ({
+      questionnaireData: {
+        context: {
+          appointment: {
+            attributes: {
+              clinicFriendlyName,
+              vdsAppointments: [{ clinic }],
+            },
+          },
+        },
+      },
+    }),
+    subscribe: () => {},
+    dispatch: () => ({}),
+  };
+};
+
+export { createFakeFooterStore };
